requirements.txt
langgraph
langchain
fastapi 
uvicorn 
sqlalchemy 
psycopg2-binary 
pymongo 
redis 
boto3
python-dotenv 
pydantic
pymongo
langchain_openai
tavily